refactor: certify runtime and separate preflight checks
