+++ src/evolve/preflight/__init__.py
@@ -1,0 +1,25 @@
+from . import checks
+from .checks import artifact_reference
+from .models import (
+    ArtifactReferenceV1,
+    PreflightCheckStatus,
+    PreflightCheckV1,
+    PreflightFailureCategory,
+    PreflightMode,
+    PreflightResultV1,
+    PreflightStatus,
+)
+from .runner import run_preflight
+
+__all__ = [
+    "ArtifactReferenceV1",
+    "PreflightCheckStatus",
+    "PreflightCheckV1",
+    "PreflightFailureCategory",
+    "PreflightMode",
+    "PreflightResultV1",
+    "PreflightStatus",
+    "artifact_reference",
+    "checks",
+    "run_preflight",
+]
